# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/exports/issues.csv": {
      "get": {
        "operationId": "ExportsController_issues",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "UTF-8 BOM 이슈 CSV 스트림"
          },
          "401": {
            "description": "SESSION_REQUIRED",
            "content": {
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "현재 워크스페이스 이슈 CSV 내보내기",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/projects.csv": {
      "get": {
        "operationId": "ExportsController_projects",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "UTF-8 BOM 프로젝트 CSV 스트림"
          },
          "401": {
            "description": "SESSION_REQUIRED",
            "content": {
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "현재 워크스페이스 프로젝트 CSV 내보내기",
        "tags": [
          "exports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rivet API",
    "description": "Rivet 비공개 MVP 베타 REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-rivet_session"
      }
    },
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "입력값을 확인해 주세요."
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "example": {
              "email": [
                "올바른 이메일 주소를 입력해 주세요."
              ]
            }
          },
          "requestId": {
            "type": "string",
            "example": "req_550e8400-e29b-41d4-a716-446655440000"
          },
          "currentVersion": {
            "type": "number",
            "minimum": 1
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ]
      },
...
    }
  }
}